Convert `integration.Sink` to use `handler.UnitOfWork`.
