+++ handler/handler.go
@@ -1,0 +1,13 @@
+package handler
+
+import (
+	"context"
+
+	"github.com/dogmatiq/infix/parcel"
+)
+
+// Handler is an interface for handling messages.
+type Handler interface {
+	// Handle handles the message in p.
+	HandleMessage(ctx context.Context, w UnitOfWork, p *parcel.Parcel) error
+}
