+++ handler/integration/doc.go
@@ -1,3 +1,3 @@
-// Package integration implements a pipeline sink that dispatches to Dogma
+// Package integration implements a handler that dispatches to Dogma
 // integration message handlers.
 package integration
